Added retry 3 times to build docker images

+++ docker/msdp/build_all.sh
@@ -2,6 +2,21 @@
 
 echo "Build all MS DeepSpeed flavoured PyTorch-Ignite images"
 
+retry()
+{
+    cmd=$1 msg=$2
+    counter=0 limit=3
+    while [ "$counter" -lt "$limit" ]; do
+        echo "(Re-)Try: $cmd"
+        bash -c "$cmd" && break
+        echo $msg
+        counter="$(( $counter + 1 ))"
+    done
+    if [ $counter -eq $limit ]; then
+        exit 1
+    fi
+}
+
 # Start script from ignite docker folder
 if [ ! -d msdp ]; then
     echo "Can not find 'msdp' folder"
@@ -12,18 +27,17 @@ fi
 curr_dir=$PWD
 cd $curr_dir/msdp
 
-set -xeu
+set -eu
 
-image_name="msdp-apex-base"
+image_tag=""
 
-docker build -t pytorchignite/${image_name}:latest -f Dockerfile.${image_name} .
-image_tag=`docker run --rm -i pytorchignite/${image_name}:latest python -c "import torch; import ignite; print(torch.__version__ + \"-\" + ignite.__version__, end=\"\")"`
-docker tag pytorchignite/${image_name}:latest pytorchignite/${image_name}:${image_tag}
-
-for image_name in "msdp-apex-vision" "msdp-apex-nlp"
+for image_name in "msdp-apex-base" "msdp-apex-vision" "msdp-apex-nlp"
 do
 
-    docker build -t pytorchignite/${image_name}:latest -f Dockerfile.${image_name} .
+    retry "docker build -t pytorchignite/${image_name}:latest -f Dockerfile.${image_name} ." "\nBuild failed: ${image_name}"
+    if [ -z $image_tag ]; then
+        image_tag=`docker run --rm -i pytorchignite/${image_name}:latest python -c "import torch; import ignite; print(torch.__version__ + \"-\" + ignite.__version__, end=\"\")"`
+    fi
     docker tag pytorchignite/${image_name}:latest pytorchignite/${image_name}:${image_tag}
 
 done
